Moved to correct directory. Hope it'll work out :)
