+++ Lib/Controllers/Roblox.Points/v1/universes/_P-universeId/users/_P-userId/all-time.d.ts
@@ -1,0 +1,5 @@
+declare const _default: {
+    method: string;
+    func: (_req: any, res: any) => Promise<any>;
+};
+export default _default;
