+++ Roblox/Helpers/WebHelpers/Points/Get.ts
@@ -1,5 +1,5 @@
 import fs from 'fs';
-import { _dirname } from '../../Constants/Directories';
+import { _dirname } from '../../../../Lib/Helpers/Constants/Directories';
 //import Base64 from 'crypto-js/enc-base64';
 //import crpto from 'crypto-js';
